requirements.txt
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
cycler==0.11.0
filelock==3.9.0
fonttools==4.38.0
ftfy==6.1.1
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
open-clip-torch==2.7.0
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.12.1
